backend/requirements.txt
flask
opencv-python
pyzbar
numpy
Pillow
